rename camelCase file name to dash-split file name
