+++ fields/types/html/image/image-editing-block.js
